requirements.txt
streamlit
pandas
python-dotenv
langchain-groq
langchain-core
langchain-experimental
langchain-community
ddgs
tabulate
matplotlib